+++ src/System.ObjectModel/tests/KeyedCollection/Serialization.cs
