+++ src/System.ObjectModel/tests/ObservableCollection/ObservableCollection_ConstructorAndPropertyTests.cs
@@ -132,5 +132,36 @@ namespace System.Collections.ObjectModel.Tests
 
             public List<T> InnerList => (List<T>)base.Items;
         }
+
+        /// <summary>
+        /// Tests that ArgumentNullException is thrown when given a null IEnumerable.
+        /// </summary>
+        [Fact]
+        public static void ListConstructorTest_Negative()
+        {
+            Assert.Throws<ArgumentNullException>("list", () => new ObservableCollection<string>((List<string>)null));
+        }
+
+        [Fact]
+        public static void ListConstructorTest()
+        {
+            List<string> collection = new List<string> { "one", "two", "three" };
+            var actual = new ObservableCollection<string>(collection);
+            Assert.Equal(collection, actual);
+        }
+
+        [Fact]
+        public static void ListConstructorTest_MakesCopy()
+        {
+            List<string> collection = new List<string> { "one", "two", "three" };
+            var oc = new ObservableCollectionSubclass<string>(collection);
+            Assert.NotNull(oc.InnerList);
+            Assert.NotSame(collection, oc.InnerList);
+        }
+
+        private partial class ObservableCollectionSubclass<T> : ObservableCollection<T>
+        {
+            public ObservableCollectionSubclass(List<T> list) : base(list) { }
+        }
     }
 }
